moved mediator specific classes in data.ef to a separate folder
