+++ src/BoltOn.Data.EF/BoltOn.Data.EF.csproj
@@ -10,4 +10,7 @@
   <ItemGroup>
     <ProjectReference Include="..\BoltOn\BoltOn.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Mediator\" />
+  </ItemGroup>
 </Project>
